requirements.txt
httpx
python-dotenv
pydantic-settings